+++ django/pokemon/urls.py
@@ -4,5 +4,7 @@ from . import views
 
 urlpatterns = [
     path('', views.index),
-    path('create', views.create)
+    path('<int:pokemon_id>', views.detail),
+    path('<int:pokemon_id>/give-xp', views.give_xp),
+    path('create', views.create),
 ]